+++ src/main/java/cc/polyfrost/oneconfig/gui/elements/Slider.java
@@ -32,7 +32,6 @@ public class Slider extends BasicElement {
 
     }
 
-    @Override
     public void update(float x, float y) {
         super.update(x, y);
         boolean isMouseDown = Platform.getMousePlatform().isButtonDown(0);
